CREATE2 Factory Deployment Dapp
This PR creates a small Dapp on a static HTML page for deploying the
CREATE2 factory contract.

**Unfortunately, it does not work**. In particular, MetaMask does not
currently support EIP-7702 type `0x4` transaction with the
`authorizationList` field, meaning that we cannot execute the
bootstrapping transaction on the deployer account. In fact, MetaMask
produces the following error:

```
External EIP-7702 transactions are not supported
```

I'm creating this as a draft PR for now, as I hope that this will
change over time (and MetaMask will allow EIP-7702 transactions).