Change The Standardised Salt
This PR proposes a new standardized salt for the CREATE2 factory that
was mined in a very deterministic way (the first one to match our
pattern starting from 0). This is to prevent any concerns that we, the
ERC authors, have mined a CREATE2 collision on the CREATE2 factory
address that we can exploit at some point in the future.